Completed validate input
